+++ test-app/webroot/TestComp.py
@@ -1,0 +1,2 @@
+def render():
+    return {"var": "really nice magic"}